% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liDoseResponseCurves.R
\name{liDoseResponseCurves}
\alias{liDoseResponseCurves}
\title{Create dose response curves UI.}
\usage{
liDoseResponseCurves(anchorId)
}
\arguments{
\item{anchorId}{A unique id the identifies the anchor.}
}
\description{
\code{liDoseResponseCurves} creates an anchor point in your UI definition. This anchor point
is where alerts created in your Server logic will be displayed.
}
\details{
See \link{liDoseResponseCurves} for more information about how to use \code{liDoseResponseCurves} with the
rest of the liDoseResponseCurves family.
}
\note{
Run \code{liExample("liDoseResponseCurves")} for an example
of \code{liDoseResponseCurves} functionality.
}
\seealso{
\href{http://}{HMS}

Other res: \code{\link{liDoseResponseGrid}},
  \code{\link{liScatterplot}},
  \code{\link{renderLiDoseResponseCurves}},
  \code{\link{renderLiDoseResponseGrid}},
  \code{\link{renderLiScatterplot}}
}

