% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liDoseResponseGrid.R
\name{liDoseResponseGrid}
\alias{liDoseResponseGrid}
\title{Create dose response grid UI.}
\usage{
liDoseResponseGrid(anchorId)
}
\arguments{
\item{anchorId}{A unique id the identifies the anchor.}
}
\description{
\code{liDoseResponseGrid} creates an anchor point in your UI definition. This anchor point
is where alerts created in your Server logic will be displayed.
}
\details{
See \link{liDoseResponseGrid} for more information about how to use \code{liDoseResponseGrid} with the
rest of the liDoseResponseGrid family.
}
\note{
Run \code{liExample("liDoseResponseGrid")} for an example
of \code{liDoseResponseGrid} functionality.
}
\seealso{
\href{http://}{HMS}

Other res: \code{\link{liDoseResponseCurves}},
  \code{\link{liScatterplot}},
  \code{\link{renderLiDoseResponseCurves}},
  \code{\link{renderLiDoseResponseGrid}},
  \code{\link{renderLiScatterplot}}
}

