% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liExample.R
\name{liExample}
\alias{liExample}
\title{Run embedded examples.}
\usage{
liExample(family, display.mode = "showcase", ...)
}
\arguments{
\item{family}{A shinyLi family name}

\item{display.mode}{The display mode to use when running the example. See
\code{\link{runApp}}.}

\item{\dots}{Other parameters to pass to \code{\link{runApp}}.}
}
\description{
A function to view examples of shinyLi functionality. Will run the examples
found in the examples sections of shinyLi documentation. Use this instead of 
\code{example}.
}
\details{
This function is just a wrapper for \code{\link{runApp}} that runs copies of the
examples found in the family documention pages of \code{shinyLi}. By default,
\code{display.mode} is set to \code{showcase} so you can see the code while 
the app is running.
}
\examples{
\dontrun{
   liExample("scatterplot1")}
}

