% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liScatterplot.R
\name{liScatterplot}
\alias{liScatterplot}
\title{Create scatterplot UI.}
\usage{
liScatterplot(anchorId)
}
\arguments{
\item{anchorId}{A unique id the identifies the anchor.}
}
\description{
\code{liScatterplot} creates an anchor point in your UI definition. This anchor point
is where alerts created in your Server logic will be displayed.
}
\details{
See \link{liScatterplot} for more information about how to use \code{liScatterplot} with the
rest of the liScatterplot family.
}
\note{
Run \code{liExample("liScatterplot")} for an example
of \code{liScatterplot} functionality.
}
\seealso{
\href{http://}{HMS}

Other res: \code{\link{liDoseResponseCurves}},
  \code{\link{liDoseResponseGrid}},
  \code{\link{renderLiDoseResponseCurves}},
  \code{\link{renderLiDoseResponseGrid}},
  \code{\link{renderLiScatterplot}}
}

