% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderLiDoseResponseCurves.R
\name{renderLiDoseResponseCurves}
\alias{renderLiDoseResponseCurves}
\title{Render dose response curves.}
\usage{
renderLiDoseResponseCurves(input)
}
\arguments{
\item{input}{A unique id the identifies the anchor.}
}
\description{
\code{renderLiDoseResponseCurves}
}
\details{
See \link{renderLiDoseResponseCurves} for more information about how to use \code{renderLiDoseResponseCurves} with the
rest of the renderLiDoseResponseCurves family.
}
\note{
Run \code{liExample("renderLiDoseResponseCurves")} for an example
of \code{renderLiDoseResponseCurves} functionality.
}
\seealso{
\href{http://}{HMS}

Other res: \code{\link{liDoseResponseCurves}},
  \code{\link{liDoseResponseGrid}},
  \code{\link{liScatterplot}},
  \code{\link{renderLiDoseResponseGrid}},
  \code{\link{renderLiScatterplot}}
}

