% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderLiDoseResponseGrid.R
\name{renderLiDoseResponseGrid}
\alias{renderLiDoseResponseGrid}
\title{Render dose response grid}
\usage{
renderLiDoseResponseGrid(data = NULL, input = NULL, xmin = 0, xmax = 0,
  factors = NULL, toggle = NULL)
}
\arguments{
\item{data}{data to render (if input filename is blank)}

\item{input}{filename to load}

\item{xmin}{x axis min}

\item{xmax}{x axis max}
}
\description{
\code{renderLiDoseResponseGrid}
}
\details{
See \link{renderLiDoseResponseGrid} for more information about how to use \code{renderLiDoseResponseGrid} with the
rest of the renderLiDoseResponseGrid family.
}
\note{
Run \code{liExample("renderLiDoseResponseGrid")} for an example
of \code{renderLiDoseResponseGrid} functionality.
}
\seealso{
\href{http://}{HMS}

Other res: \code{\link{liDoseResponseCurves}},
  \code{\link{liDoseResponseGrid}},
  \code{\link{liScatterplot}},
  \code{\link{renderLiDoseResponseCurves}},
  \code{\link{renderLiScatterplot}}
}

