% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderLiScatterplot.R
\name{renderLiScatterplot}
\alias{renderLiScatterplot}
\title{Render scatterplot.}
\usage{
renderLiScatterplot(data = NULL, input = NULL, factors = NULL)
}
\arguments{
\item{input}{A unique id the identifies the anchor.}
}
\description{
\code{renderLiScatterplot}
}
\details{
See \link{renderLiScatterplot} for more information about how to use \code{renderLiScatterplot} with the
rest of the renderLiScatterplot family.
}
\note{
Run \code{liExample("renderLiScatterplot")} for an example
of \code{renderLiScatterplot} functionality.
}
\seealso{
\href{http://}{HMS}

Other res: \code{\link{liDoseResponseCurves}},
  \code{\link{liDoseResponseGrid}},
  \code{\link{liScatterplot}},
  \code{\link{renderLiDoseResponseCurves}},
  \code{\link{renderLiDoseResponseGrid}}
}

